refactor : extract calculatorReducer and inject dependencies
