+++ Task1/Task1/ContentView.swift
@@ -11,64 +11,22 @@ import ComposableArchitecture
 struct ContentView: View {
     let store: Store<CalculatorState, CalculatorAction>
     
-    let calculatorReducer : AnyReducer<CalculatorState, CalculatorAction, CalculatorEnvironment>
-    
-    
-    init(environment: CalculatorEnvironment) {
-        calculatorReducer = AnyReducer<CalculatorState, CalculatorAction, CalculatorEnvironment> {state, action, environment in switch action {
-        case let .firstNumberChanged(number):
-            state.firstNumber = number
-            return .none
-            
-        case let .secondNumberChanged(number):
-            state.secondNumber = number
-            return .none
-            
-        case .add:
-            if let first = Int(state.firstNumber), let second = Int(state.secondNumber) {
-                let result = environment.calculatorService.add(first, second)
-                state.resultString = "\(first) + \(second) = \(result)"
-            } else {
-                state.resultString = "잘못된 입력입니다"
-            }
-            return .none
-        case .sub:
-            if let first = Int(state.firstNumber), let second = Int(state.secondNumber) {
-                let result = environment.calculatorService.subtract(first, second)
-                state.resultString = "\(first) - \(second) = \(result)"
-            } else {
-                state.resultString = "잘못된 입력입니다"
-            }
-            return .none
-        case .multi:
-            if let first = Int(state.firstNumber), let second = Int(state.secondNumber) {
-                let result = environment.calculatorService.multiply(first, second)
-                state.resultString = "\(first) * \(second) = \(result)"
-            } else {
-                state.resultString = "잘못된 입력입니다"
-            }
-            return .none
-        case .div:
-            if let first = Int(state.firstNumber), let second = Int(state.secondNumber) {
-                if second == 0 {
-                    state.resultString = "0으로 나눌 수 없습니다!"
-                } else {
-                    let result = environment.calculatorService.divide(first, second)
-                    state.resultString = "\(first) / \(second) = \(result)"
-                }
-                
-            } else {
-                state.resultString = "잘못된 입력입니다"
-            }
-            return .none
-        }}
-        
-        self.store = Store(initialState: CalculatorState(), reducer: calculatorReducer, environment: environment)
-        
+    init(store: Store<CalculatorState, CalculatorAction>) {
+        self.store = store
     }
     
     var body: some View {
         CalculatorView(store: store)
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        let calculatorService = CalculatorServiceImpl()
+        let environment = CalculatorEnvironment(calculatorService: calculatorService)
+        
+        let store = Store(initialState: CalculatorState(), reducer: calculatorReducer, environment: environment)
+        return ContentView(store: store)
     }
 }
 
@@ -179,12 +137,5 @@ extension Button where Label == Text {
             .padding()
             .frame(width: 300, height: 40)
             .background(RoundedRectangle(cornerRadius: 30).fill(Color.teal))
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        let environment = CalculatorEnvironment(calculatorService: CalculatorServiceImpl())
-        return ContentView(environment: environment)
     }
 }
